:recycle: refactor: move core package structure under room domain
