+++ kok-api/src/main/java/com/kok/kokapi/config/usecase/UseCaseConfig.java
@@ -1,8 +1,8 @@
 package com.kok.kokapi.config.usecase;
 
 import com.kok.kokapi.room.application.service.RoomCommandService;
-import com.kok.kokcore.application.port.out.SaveRoomPort;
-import com.kok.kokcore.usecase.CreateRoomUseCase;
+import com.kok.kokcore.room.application.port.out.SaveRoomPort;
+import com.kok.kokcore.room.usecase.CreateRoomUseCase;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
