+++ kok-core/build.gradle
@@ -19,8 +19,6 @@ repositories {
 
 dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter'
-	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
-	implementation 'com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.18.2'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
 }
