+++ kok-core/src/main/java/com/kok/kokcore/room/usecase/GetRoomUseCase.java
@@ -1,0 +1,7 @@
+package com.kok.kokcore.room.usecase;
+
+import com.kok.kokcore.room.domain.Room;
+
+public interface GetRoomUseCase {
+    Room getRoomById(String roomId);
+}
