--- kok-core/src/main/java/com/kok/kokcore/usecase/GetRoomUseCase.java
@@ -1,7 +1,0 @@
-package com.kok.kokcore.usecase;
-
-import com.kok.kokcore.domain.Room;
-
-public interface GetRoomUseCase {
-    Room getRoomById(String roomId);
-}
